pass a badge as prop
