+++ packages/gamut-labs/src/AppHeader/AppHeaderElements/AppHeaderLinkSections/index.tsx
@@ -1,3 +1,5 @@
+import { Badge } from '@codecademy/gamut';
+import styled from '@emotion/styled';
 import React from 'react';
 
 import { AppHeaderLinkMobile } from '../../../AppHeaderMobile/AppHeaderLinkMobile';
@@ -11,6 +13,16 @@ export type AppHeaderLinkSectionsProps = {
   action: AppHeaderClickHandler;
   item: AppHeaderDropdownItem;
 };
+
+export const StyledBadge = styled(Badge)`
+  font-size: 10px;
+  padding-top: 1px;
+  padding-bottom: 10px;
+  padding-left: 8px;
+  padding-right: 8px;
+  width: 36px;
+  border-radius: 16px;
+`;
 
 export const AppHeaderLinkSections: React.FC<AppHeaderLinkSectionsProps> = ({
   action,
@@ -27,7 +39,22 @@ export const AppHeaderLinkSections: React.FC<AppHeaderLinkSectionsProps> = ({
                   item={link}
                   key={link.id}
                   topSeparator={sectionIndex !== 0 && linkIndex === 0}
-                  newBadge={link.newBadge}
+                  badge={
+                    link.newBadge && (
+                      <StyledBadge
+                        round
+                        variant="blue"
+                        fontFamily="accent"
+                        ml={8}
+                        pl={8}
+                        pr={8}
+                        height={16}
+                        alignSelf="center"
+                      >
+                        New
+                      </StyledBadge>
+                    )
+                  }
                 />
               );
             });
@@ -38,7 +65,22 @@ export const AppHeaderLinkSections: React.FC<AppHeaderLinkSectionsProps> = ({
                 action={action}
                 item={link}
                 key={link.id}
-                newBadge={link.newBadge}
+                badge={
+                  link.newBadge && (
+                    <StyledBadge
+                      round
+                      variant="blue"
+                      fontFamily="accent"
+                      ml={8}
+                      pl={8}
+                      pr={8}
+                      height={16}
+                      alignSelf="center"
+                    >
+                      New
+                    </StyledBadge>
+                  )
+                }
               />
             );
           })}
